requirements.txt
asgiref==3.8.1
cx-Oracle==8.3.0
Django==5.0.9
django-cors-headers==4.5.0
django-mssql-backend==2.8.1
djangorestframework==3.15.2
mssql-django==1.5
mysqlclient==2.2.4
psycopg2==2.9.9
pyodbc==5.1.0
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
